1. 加入鬧鐘可以更改時間功能（下修倒數時間功能） 2. 修正建立/呼叫鬧鐘時，不再判斷 0 秒倒數問題，交由倒數計時器直接 reset

+++ TestColckManager/TestColckManager/ViewController.m
@@ -35,7 +35,7 @@
             buttonTitle = @"開始";
         }
         else{
-            buttonTitle = [NSString stringWithFormat:@"倒數 %2lu 秒！" ,second];
+            buttonTitle = [NSString stringWithFormat:@"倒數 %2lu 秒！" ,(unsigned long)second];
         }
         [strongSelf->_testButton setTitle:buttonTitle forState:UIControlStateNormal];
     }];
@@ -65,7 +65,7 @@
             buttonTitle = @"開始";
         }
         else{
-            buttonTitle = [NSString stringWithFormat:@"倒數 %2lu 秒！" ,second];
+            buttonTitle = [NSString stringWithFormat:@"倒數 %2lu 秒！" ,(unsigned long)second];
         }
         [strongSelf->_testButton setTitle:buttonTitle forState:UIControlStateNormal];
     }];
